moving PrescaleTableModel into its own file
